[UPDATE] Move all models to models package
